Optimized interactive extraction by reducing the triggers to figure-changing callbacks; introduced a binning option, which downgrades the figure size, further improving performances; improved layout disposition; improved value errors handling, making a dedicated div to report the error to the user and making the error not app-crashing; introduce server-side output for some specific components; improved chache handling, with automatic cache clearing at the start and end of application.

+++ GONet_Wizard/GONet_utils/src/extractors.py
@@ -86,7 +86,7 @@ class extraction_output:
         Number of pixels within the circular region.
 
     """
-    total_counts: float
+    total_counts: int
     mean_counts: float
     std: float
     npixels: int
